couple of fixes to compiles on windows, they aren't proper fixes but nobody answered on the windows ml, at least people can compile kdelibs :)
svn path=/trunk/KDE/kdelibs/; revision=972538

+++ experimental/knotificationitem/knotificationitemdbus_p.cpp
@@ -195,7 +195,7 @@ QString KNotificationItemDBus::Status() const
 int KNotificationItemDBus::WindowId() const
 {
     if (m_notificationItem->d->associatedWidget) {
-        return m_notificationItem->d->associatedWidget->winId();
+        return (int)m_notificationItem->d->associatedWidget->winId();
     } else {
         return 0;
     }
